Setup pretty URL for web (#151)
To do this, I had to create an ACM certificate in us-east-1 for the new domain. This is due to that CloudFront require ACM certificates to exist in us-east-1.
In order to accomplish this, I introduced a new stack separate from the 'main stack,', which is deployed to us-east-1.
The ARN of the ACM certificate is put into ParameterStore, and then obtained in the main stack to configure CloudFront.